Issues with travis CI fixes
