+++ .travis.yml
@@ -14,3 +14,4 @@ script:
   - py.test tests/test_ecvalidatingfiles.py -vv
   - py.test tests/test_ecprediction.py -vv
   - py.test tests/test_virulencefactors.py -vv
+  - py.test tests/test_rgi.py -vv
